refactor: simplify foundation architecture and improve naming
- Merge PromptManager into ResourceManager to eliminate unnecessary abstraction
- Rename base_managers.py → resource_manager.py to match class name
- Rename interfaces.py → enums.py to accurately reflect enum contents
- Update all imports throughout codebase consistently
- Maintain all functionality while simplifying inheritance hierarchy

This refactoring reduces cognitive load and follows Python naming conventions better.
