+++ ray_mcp/kubernetes/kuberay_job_manager.py
@@ -3,8 +3,8 @@
 from typing import Any, Dict, Optional
 
 from ..config import get_config_manager_sync
-from ..foundation.base_managers import ResourceManager
 from ..foundation.logging_utils import LoggingUtility
+from ..foundation.resource_manager import ResourceManager
 from ..parsers import ActionParser
 from .manifest_generator import ManifestGenerator
