+++ ray_mcp/managers/cluster_manager.py
@@ -8,7 +8,7 @@ import subprocess
 from typing import Any, Dict, List, Optional, Tuple
 
 from ..config import get_config_manager_sync
-from ..foundation.base_managers import ResourceManager
+from ..foundation.resource_manager import ResourceManager
 from ..parsers import ActionParser
