+++ ray_mcp/managers/job_manager.py
@@ -4,7 +4,7 @@ import asyncio
 from typing import Any, Dict
 
 from ..config import get_config_manager_sync
-from ..foundation.base_managers import ResourceManager
+from ..foundation.resource_manager import ResourceManager
 from ..parsers import ActionParser
